feat(core): extract paths + run-registry (KJC-TSK-0511)
PR3 of the KJC-TSK-0511 monorepo decoupling: lift the HOME-path resolver
and the run-registry persistence layer into @karajan/core so the
hu-board workspace can consume them without reaching into src/utils/.

Originals become 8-line shims that re-export from the core subpaths,
keeping every CLI caller untouched. The hu-board run-tracker now imports
listActiveRuns from @karajan/core/run-registry directly. New smoke
tests (paths, run-registry) on top of PR2's coverage take the core
workspace to 15 tests across 5 files.
