+++ packages/hu-board/src/run-tracker.js
@@ -16,7 +16,7 @@
  * se borran como cleanup (responsabilidad de run-registry).
  */
 
-import { listActiveRuns as fsList } from "../../../src/utils/run-registry.js";
+import { listActiveRuns as fsList } from "@karajan/core/run-registry";
 
 const runs = new Map();
 let reaperTimer = null;
